backend: get poops with location for user, lat & long now float (was int), returning types rather than always poop class (cleaner and simple), added some dummy data to feed.

+++ back-end/repository/prisma/migrations/20241116204141_init/migration.sql
@@ -36,8 +36,8 @@ CREATE TABLE "Poop" (
     "colorID" INTEGER,
     "rating" INTEGER NOT NULL,
     "title" TEXT,
-    "latitude" INTEGER,
-    "longitude" INTEGER,
+    "latitude" DOUBLE PRECISION,
+    "longitude" DOUBLE PRECISION,
     "userID" INTEGER NOT NULL,
 
     CONSTRAINT "Poop_pkey" PRIMARY KEY ("poopID")
